requirements.txt
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
prometheus-client
